Refactor use cases to return output data instead of callback

+++ src/test/java/ca/axoplasm/Octoscribe/use_case/createSubtitledVideo/CreateSubtitledVideoInteractorTest.java
@@ -1,7 +1,6 @@
 package ca.axoplasm.Octoscribe.use_case.createSubtitledVideo;
 
 import ca.axoplasm.Octoscribe.data_access.MediaConvertObject;
-import ca.axoplasm.Octoscribe.use_case.videoToAudio.*;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -10,62 +9,38 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class CreateSubtitledVideoInteractorTest {
     @Test
-    public void successtest() {
+    public void successTest() {
         File video = new File("src/test/resources/TestVideo.mp4");
         File subtitle = new File("src/test/resources/TestVideo.srt");
         CreateSubtitledVideoInputData input = new CreateSubtitledVideoInputData(video, subtitle);
         MediaConvertObject mci = new MediaConvertObject();
 
-        CreateSubtitledVideoOutputBoundary boundary = new CreateSubtitledVideoOutputBoundary() {
-            @Override
-            public void prepareSuccessView(CreateSubtitledVideoOutputData data) {
-                File sVideo = new File("src/test/resources/subtitledTestVideo.mp4");
-                try {
-                    assertEquals("src/test/resources/subtitledTestVideo.mp4", data.getFileName());
-                    assertFalse(data.getUseCaseFailed());
-                    sVideo.delete();
-                } catch (Exception e) {
-                    sVideo.delete();
-                    assert false;
-                }
-            }
+        CreateSubtitledVideoInputBoundary inputBoundary = new CreateSubtitledVideoInteractor(mci);
+        CreateSubtitledVideoOutputData output = inputBoundary.execute(input);
+        assertFalse(output.getUseCaseFailed());
 
-            @Override
-            public void prepareFailureView(String errorMessage) {
-                fail("Use case failure is unexpected.");
+        if (!output.getUseCaseFailed()) {
+            File sVideo = new File("src/test/resources/subtitledTestVideo.mp4");
+            try {
+                assertEquals("src/test/resources/subtitledTestVideo.mp4", output.getFileName());
+                sVideo.delete();
+            } catch (Exception e) {
+                sVideo.delete();
+                assert false;
             }
-        };
-
-        CreateSubtitledVideoInputBoundary inputBoundary = new CreateSubtitledVideoInteractor(mci, boundary);
-        inputBoundary.execute(input);
+        }
     }
 
     @Test
-    public void failuretest() {
+    public void failureTest() {
         File video = new File("src/test/resources/notreal.mp4");
         File subtitle = new File("src/test/resources/alsonotreal.srt");
         CreateSubtitledVideoInputData input = new CreateSubtitledVideoInputData(video, subtitle);
         MediaConvertObject mci = new MediaConvertObject();
 
-        CreateSubtitledVideoOutputBoundary boundary = new CreateSubtitledVideoOutputBoundary() {
-            @Override
-            public void prepareSuccessView(CreateSubtitledVideoOutputData data) {
-                try {
-                    fail("Use case success is unexpected.");
-                } catch (Exception e) {
-                    File audio = new File("src/test/resources/subtitledTestVideo.mp4");
-                    audio.delete();
-                }
-            }
-
-            @Override
-            public void prepareFailureView(String errorMessage) {
-                assertEquals("Video Conversion Failed", errorMessage);
-            }
-        };
-
-        CreateSubtitledVideoInputBoundary inputBoundary = new CreateSubtitledVideoInteractor(mci, boundary);
-        inputBoundary.execute(input);
+        CreateSubtitledVideoInputBoundary inputBoundary = new CreateSubtitledVideoInteractor(mci);
+        CreateSubtitledVideoOutputData output = inputBoundary.execute(input);
+        assert output.getUseCaseFailed();
     }
 
     @Test
@@ -75,26 +50,19 @@ class CreateSubtitledVideoInteractorTest {
         CreateSubtitledVideoInputData input = new CreateSubtitledVideoInputData(video, subtitle);
         MediaConvertObject mci = new MediaConvertObject();
 
-        CreateSubtitledVideoOutputBoundary boundary = new CreateSubtitledVideoOutputBoundary() {
-            @Override
-            public void prepareSuccessView(CreateSubtitledVideoOutputData data) {
-                File svideo = new File("src/test/resources/subtitledTestVideo.mp4");
-                try {
-                    assert svideo.exists();
-                    svideo.delete();
-                } catch (Exception e) {
-                    svideo.delete();
-                    assert false;
-                }
-            }
+        CreateSubtitledVideoInputBoundary inputBoundary = new CreateSubtitledVideoInteractor(mci);
+        CreateSubtitledVideoOutputData output = inputBoundary.execute(input);
+        assertFalse(output.getUseCaseFailed());
 
-            @Override
-            public void prepareFailureView(String errorMessage) {
-                fail("Use case failure is unexpected.");
+        if (!output.getUseCaseFailed()) {
+            File svideo = new File("src/test/resources/subtitledTestVideo.mp4");
+            try {
+                assert svideo.exists();
+                svideo.delete();
+            } catch (Exception e) {
+                svideo.delete();
+                assert false;
             }
-        };
-
-        CreateSubtitledVideoInputBoundary inputBoundary = new CreateSubtitledVideoInteractor(mci, boundary);
-        inputBoundary.execute(input);
+        }
     }
 }
